requirements.txt
asgiref==3.7.2
autopep8==2.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==1.2.0
Django==4.2.3
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
packaging==23.1
pipreqs==0.4.13
psycopg2==2.9.5
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarg==0.1.9